requirements.txt
feedparser==6.0.11
requests==2.32.3
beautifulsoup4==4.12.3
pyyaml==6.0.2
jinja2==3.1.4
python-dateutil==2.9.0
